Refactor geo testing
